Review feedback part 1:
split token-login stuff into it's own service
user-edit radios
user-edit timeout tests wrapped in try-catch
padding instead of calc
