+++ api/src/public/login/style.css
@@ -43,8 +43,8 @@ input {
   margin: 2em auto;
   font-size: 1em;
   width: 302px;
-  max-width: calc(100% - 20px);
-  padding: 6px 10px;
+  max-width: 100%;
+  padding: 6px 0;
   text-decoration: none;
 }
 .loading .btn {
