+++ api/src/services/users.js
@@ -6,15 +6,11 @@ const db = require('../db');
 const lineage = require('@medic/lineage')(Promise, db.medic);
 const getRoles = require('./types-and-roles');
 const auth = require('../auth');
-const config = require('../config');
-const phoneNumber = require('@medic/phone-number');
-const taskUtils = require('@medic/task-utils');
+const tokenLogin = require('./token-login');
 
 const USER_PREFIX = 'org.couchdb.user:';
 const ONLINE_ROLE = 'mm-online';
 const DOC_IDS_WARN_LIMIT = 10000;
-
-const LOGIN_TOKEN_EXPIRE_TIME = 24 * 60 * 60 * 1000; // 24 hours
 
 const PASSWORD_MINIMUM_LENGTH = 8;
 const PASSWORD_MINIMUM_SCORE = 50;
@@ -63,21 +59,6 @@ const illegalDataModificationAttempts = data =>
 const error400 = (msg, key, params) => ({
   code: 400, message: { message: msg, translationKey: key, translationParams: params }
 });
-
-/**
- * Generates a complex enough password with a given length
- * @param {Number} length
- * @returns {string} - the generated password
- */
-const generatePassword = (length = PASSWORD_MINIMUM_LENGTH) => {
-  const chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789.,:!$-=';
-  let password;
-  do {
-    password = Array.from({ length }).map(() => _.sample(chars)).join('');
-  } while (passwordTester(password) < PASSWORD_MINIMUM_SCORE);
-
-  return password;
-};
 
 const getType = user => {
   if (user.roles && user.roles.length) {
@@ -437,29 +418,10 @@ const validatePassword = (password) => {
   }
 };
 
-/**
- * Given user data, validates whether the phone field is filled with a valid phone number.
- * Throws an error if the phone number field is empty or value is not valid.
- * If valid, normalizes the field value.
- * @param {Object} data - the user data
- * @param {String} data.phone - the phone number to validate
- */
-const validateAndNormalizePhone = (data) => {
-  const settings = config.get();
-  if (!phoneNumber.validate(settings, data.phone)) {
-    return error400(
-      'A valid phone number is required for SMS login.',
-      'configuration.enable.token.login.phone'
-    );
-  }
-
-  data.phone = phoneNumber.normalize(settings, data.phone);
-};
-
 const missingFields = data => {
   const required = ['username'];
 
-  if (shouldEnableTokenLogin(data)) {
+  if (tokenLogin.shouldEnableTokenLogin(data)) {
     required.push('phone');
   } else {
     required.push('password');
@@ -496,191 +458,6 @@ const getUpdatedSettingsDoc = (username, data) => {
   });
 };
 
-const validTokenLoginConfig = () => {
-  const tokenLoginConfig = config.get('token_login');
-  const isValidConfig = tokenLoginConfig &&
-                        tokenLoginConfig.enabled &&
-                        (tokenLoginConfig.translation_key || tokenLoginConfig.message) &&
-                        tokenLoginConfig.app_url;
-  return isValidConfig;
-};
-
-const shouldEnableTokenLogin = data => validTokenLoginConfig() && data.token_login;
-const shouldDisableTokenLogin = data => data.token_login === false;
-
-// this is not meant to be secure, just consistent!
-const encodeUsername = username => {
-  return Buffer.from(username).toString('base64');
-};
-const decodeUsername = encoded => {
-  return Buffer.from(encoded, 'base64').toString('utf8');
-};
-
-/**
- * Clears pending tasks in the currently assigned `token_login_sms` doc
- * Used to either disable or refresh token-login
- *
- * @param {Object} user - the _users doc
- * @param {Object} user.token_login - token-login information
- * @param {Boolean} user.token_login.active - whether the token-login is active or not
- * @param {String} user.token_login.doc_id - the id of the current `token_login_sms`
- * @returns {Promise}
- */
-const clearOldLoginSms = (user) => {
-  if (!user.token_login || !user.token_login.active || !user.token_login.doc_id) {
-    return Promise.resolve();
-  }
-
-  return db.medic
-    .get(user.token_login.doc_id)
-    .then(doc => {
-      if (!doc || !doc.tasks) {
-        return;
-      }
-
-      const pendingTasks = doc.tasks.filter(task => task.state === 'pending');
-      if (!pendingTasks.length) {
-        return;
-      }
-
-      pendingTasks.forEach(task => taskUtils.setTaskState(task, 'cleared'));
-      return db.medic.put(doc);
-    })
-    .catch(err => {
-      if (err.status === 404) {
-        return;
-      }
-      throw err;
-    });
-};
-
-/**
- * Generates a doc of type `token_login_sms` that contains the two tasks required for token login: one reserved
- * for the URL and one containing instructions for the user.
- * If the user already had token-login activated, the pending tasks of the existent `token_login_sms` doc are cleared.
- * @param {Object} user - the _users document
- * @param {Object} userSettings - the user-settings document from the main database
- * @param {String} token - the randomly generated token
- * @param {Object} tokenLoginConfig - the token_login config section that contains the translation_key used to generate
- *                                    the instructions sms message
- * @returns {Promise<Object>} - returns the result of saving the new token_login_sms document
- */
-const generateLoginSms = (user, userSettings, token, tokenLoginConfig) => {
-  return clearOldLoginSms(user).then(() => {
-    const doc = {
-      type: 'token_login_sms',
-      reported_date: new Date().getTime(),
-      user: user._id,
-      tasks: [],
-    };
-    const encoded = encodeUsername(user.name);
-    const url = `${tokenLoginConfig.app_url.replace(/\/+$/, '')}/medic/login/token/${token}/${encoded}`;
-
-    const messagesLib = config.getTransitionsLib().messages;
-
-    const context = { templateContext: Object.assign({}, userSettings, user) };
-    messagesLib.addMessage(doc, tokenLoginConfig, userSettings.phone, context);
-    messagesLib.addMessage(doc, { message: url }, userSettings.phone);
-
-    return db.medic.post(doc);
-  });
-};
-
-/**
- * Enables or disables token-login for a user
- * When enabling, if `token-login` configuration is missing or invalid, no changes are made.
- * @param {Object} response - the response of previous actions
- * @returns {Promise<{Object}>} - updated response to be sent to the client
- */
-const manageTokenLogin = (data, response) => {
-  if (shouldDisableTokenLogin(data)) {
-    return disableTokenLogin(response);
-  }
-
-  if (!shouldEnableTokenLogin(data)) {
-    return Promise.resolve(response);
-  }
-
-  return enableTokenLogin(response);
-};
-
-/**
- * Enables token-login for a user
- * - a new document is created in the `medic` database that contains tasks (SMSs) to be sent to the phone number
- *   belonging to the current user (userSettings.phone) that contain instructions and a url to login in the app
- * - if the user already had token-login enabled, the previous sms document's tasks are cleared
- * - updates the user document to contain information about the active `token-login` context for the user
- * - updates the user-settings document to contain some information to be displayed in the admin page
- * @param {Object} response - the response of previous actions
- * @returns {Promise<{Object}>} - updated response to be sent to the client
- */
-const enableTokenLogin = (response) => {
-  return Promise
-    .all([
-      validateUser(response.user.id),
-      validateUserSettings(response['user-settings'].id),
-    ])
-    .then(([ user, userSettings ]) => {
-      const tokenLoginConfig = config.get('token_login');
-      const token = generatePassword(50);
-
-      return generateLoginSms(user, userSettings, token, tokenLoginConfig)
-        .then(result => {
-          user.token_login = {
-            active: true,
-            token,
-            expiration_date: new Date().getTime() + LOGIN_TOKEN_EXPIRE_TIME,
-            doc_id: result.id
-          };
-
-          userSettings.token_login = {
-            active: true,
-            expiration_date: user.token_login.expiration_date,
-          };
-
-          response.token_login = {
-            id: result.id,
-            expiration_date: user.token_login.expiration_date,
-          };
-
-          return Promise.all([ db.users.put(user), db.medic.put(userSettings) ]);
-        });
-    })
-    .then(() => response);
-};
-
-/**
- * Disables token-login for a user.
- * Deletes the `token_login` properties from the user and userSettings doc.
- * Clears pending tasks in existent SMSs
- *
- * @param {Object} response - the response of previous actions
- * @returns {Promise<{Object}>} - updated response to be sent to the client
- */
-const disableTokenLogin = (response) => {
-  return Promise
-    .all([
-      validateUser(response.user.id),
-      validateUserSettings(response['user-settings'].id),
-    ])
-    .then(([ user, userSettings ]) => {
-      if (!user.token_login) {
-        return;
-      }
-
-      return clearOldLoginSms(user).then(() => {
-        delete user.token_login;
-        delete userSettings.token_login;
-
-        return Promise.all([
-          db.medic.put(userSettings),
-          db.users.put(user),
-        ]);
-      });
-    })
-    .then(() => response);
-};
-
 /*
  * Everything not exported directly is private.  Underscore prefix is only used
  * to export functions needed for testing.
@@ -713,13 +490,10 @@ module.exports = {
         { 'fields': missing.join(', ') }
       ));
     }
-    if (shouldEnableTokenLogin(data)) {
-      // if token-login is requested, we will need to send sms-s to the user's phone number.
-      const phoneError = validateAndNormalizePhone(data);
-      if (phoneError) {
-        return Promise.reject(phoneError);
-      }
-      data.password = generatePassword(20);
+
+    const tokenLoginError = tokenLogin.validateTokenLogin(data, true);
+    if (tokenLoginError) {
+      return Promise.reject(error400(tokenLoginError.msg, tokenLoginError.key));
     }
     const passwordError = validatePassword(data.password);
     if (passwordError) {
@@ -733,7 +507,7 @@ module.exports = {
       .then(() => storeUpdatedPlace(data))
       .then(() => createUser(data, response))
       .then(() => createUserSettings(data, response))
-      .then(() => manageTokenLogin(data, response))
+      .then(() => tokenLogin.manageTokenLogin(data, response))
       .then(() => response);
   },
 
@@ -788,10 +562,6 @@ module.exports = {
       ));
     }
 
-    if (shouldEnableTokenLogin(data)) {
-      data.password = generatePassword(20);
-    }
-
     if (data.password) {
       const passwordError = validatePassword(data.password);
       if (passwordError) {
@@ -805,17 +575,9 @@ module.exports = {
         getUpdatedSettingsDoc(username, data),
       ])
       .then(([ user, settings ]) => {
-        // when disabling token login for a user that had it enabled, setting a password is required
-        if (shouldDisableTokenLogin(data) && user.token_login && !data.password) {
-          return Promise.reject(validatePassword(''));
-        }
-
-        if (shouldEnableTokenLogin(data)) {
-          // if token-login is requested, we will need to send sms-s to the user's phone number.
-          const phoneError = validateAndNormalizePhone(settings);
-          if (phoneError) {
-            return Promise.reject(phoneError);
-          }
+        const tokenLoginError = tokenLogin.validateTokenLogin(data, false, user, settings);
+        if (tokenLoginError) {
+          return Promise.reject(error400(tokenLoginError.msg, tokenLoginError.key));
         }
 
         const response = {};
@@ -869,102 +631,10 @@ module.exports = {
               rev: resp.rev
             };
           })
-          .then(() => manageTokenLogin(data, response))
+          .then(() => tokenLogin.manageTokenLogin(data, response))
           .then(() => response);
       });
   },
 
   DOC_IDS_WARN_LIMIT,
-
-  /**
-   * Searches for a user with active, not expired token-login that matches the requested token and user.
-   *
-   * @param {String} token
-   * @param {String} encryptedUsername
-   * @returns {Promise<Object>} the id of the matched user
-   */
-  getUserByToken: (token, encoded) => {
-    const invalid = { status: 401, error: 'invalid' };
-    const expired = { status: 401, error: 'expired' };
-    if (!token || !encoded) {
-      return Promise.reject(invalid);
-    }
-
-    const username = decodeUsername(encoded);
-    if (!username) {
-      return Promise.reject(invalid);
-    }
-
-    return db.users
-      .get(createID(username))
-      .then(user => {
-        if (!user.token_login || !user.token_login.active) {
-          throw invalid;
-        }
-
-        if (user.token_login.token !== token) {
-          throw invalid;
-        }
-
-        if (user.token_login.expiration_date <= new Date().getTime()) {
-          throw expired;
-        }
-
-        return user._id;
-      })
-      .catch(err => {
-        if (err.status === 404) {
-          throw invalid;
-        }
-        throw err;
-      });
-  },
-
-  /**
-   * Generates a new random password for the user. Returns this password to be used to generate a session for this user.
-   * @param userId
-   * @returns {Promise<Object>} - object containing the user's name and the new password
-   */
-  resetPassword: userId => {
-    return validateUser(userId).then(user => {
-      if (!user.token_login || !user.token_login.active) {
-        return Promise.reject({ code: 400, message: 'invalid user' });
-      }
-
-      user.password = generatePassword();
-      return db.users.put(user).then(() => ({ user: user.name, password: user.password }));
-    });
-  },
-
-  /**
-   * Updates the user and userSettings docs when the token login link is accessed successfully.
-   * @param {String} userId - the user's id to login via token link
-   * @returns {Promise}
-   */
-  deactivateTokenLogin: userId => {
-    return Promise
-      .all([
-        validateUser(userId),
-        validateUserSettings(userId),
-      ])
-      .then(([ user, userSettings ]) => {
-        if (!user.token_login || !user.token_login.active) {
-          return Promise.reject({ code: 400, message: 'invalid user' });
-        }
-
-        const updates = {
-          active: false,
-          login_date: new Date().getTime(),
-        };
-        user.token_login = Object.assign(user.token_login, updates);
-        userSettings.token_login = Object.assign(userSettings.token_login, updates);
-
-        return Promise.all([
-          db.medic.put(userSettings),
-          db.users.put(user),
-        ]);
-      });
-  },
-
-  validTokenLoginConfig,
 };
