+++ api/tests/mocha/controllers/login.spec.js
@@ -5,6 +5,7 @@ const environment = require('../../../src/environment');
 const auth = require('../../../src/auth');
 const cookie = require('../../../src/services/cookie');
 const users = require('../../../src/services/users');
+const tokenLogin = require('../../../src/services/token-login');
 const db = require('../../../src/db').medic;
 const sinon = require('sinon');
 const config = require('../../../src/config');
@@ -286,19 +287,19 @@ describe('login controller', () => {
   describe('POST login/token', () => {
     it('should redirect the user directly if they have a valid session', () => {
       sinon.stub(auth, 'getUserCtx').resolves({ name: 'user' });
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
       sinon.stub(res, 'send').returns(res);
       sinon.stub(res, 'status').returns(res);
       return controller.tokenPost(req, res).then(() => {
         chai.expect(auth.getUserCtx.callCount).to.equal(1);
         chai.expect(auth.getUserCtx.args[0]).to.deep.equal([req]);
-        chai.expect(users.validTokenLoginConfig.callCount).to.equal(0);
+        chai.expect(tokenLogin.validTokenLoginConfig.callCount).to.equal(0);
       });
     });
 
     it('should fail early when token login not enabled', () => {
       sinon.stub(auth, 'getUserCtx').rejects({ code: 401 });
-      sinon.stub(users, 'validTokenLoginConfig').returns(false);
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(false);
       sinon.stub(res, 'json').returns(res);
       sinon.stub(res, 'status').returns(res);
       return controller.tokenPost(req, res).then(() => {
@@ -311,7 +312,7 @@ describe('login controller', () => {
 
     it('should fail early with no params', () => {
       sinon.stub(auth, 'getUserCtx').rejects({ code: 401 });
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
       sinon.stub(res, 'json').returns(res);
       sinon.stub(res, 'status').returns(res);
       req.params = { token: 'my_token' };
@@ -325,14 +326,14 @@ describe('login controller', () => {
 
     it('should send 401 when token incorrect', () => {
       sinon.stub(auth, 'getUserCtx').rejects({ code: 401 });
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
-      sinon.stub(users, 'getUserByToken').resolves(false);
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'getUserByToken').resolves(false);
       sinon.stub(res, 'json').returns(res);
       sinon.stub(res, 'status').returns(res);
       req.params = { token: 'my_token', userId: 'myUserID' };
       return controller.tokenPost(req, res).then(() => {
-        chai.expect(users.getUserByToken.callCount).to.equal(1);
-        chai.expect(users.getUserByToken.args[0]).to.deep.equal( [ 'my_token', 'myUserID' ]);
+        chai.expect(tokenLogin.getUserByToken.callCount).to.equal(1);
+        chai.expect(tokenLogin.getUserByToken.args[0]).to.deep.equal( [ 'my_token', 'myUserID' ]);
         chai.expect(res.status.callCount).to.equal(1);
         chai.expect(res.status.args[0]).to.deep.equal([401]);
         chai.expect(res.json.callCount).to.equal(1);
@@ -342,13 +343,13 @@ describe('login controller', () => {
 
     it('should send error when error thrown while validating token', () => {
       sinon.stub(auth, 'getUserCtx').rejects({ code: 401 });
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
-      sinon.stub(users, 'getUserByToken').rejects({ some: 'err' });
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'getUserByToken').rejects({ some: 'err' });
       sinon.stub(res, 'json').returns(res);
       sinon.stub(res, 'status').returns(res);
       req.params = { token: 'a', userId: 'b' };
       return controller.tokenPost(req, res).then(() => {
-        chai.expect(users.getUserByToken.callCount).to.equal(1);
+        chai.expect(tokenLogin.getUserByToken.callCount).to.equal(1);
         chai.expect(res.status.callCount).to.equal(1);
         chai.expect(res.status.args[0]).to.deep.equal([400]);
         chai.expect(res.json.callCount).to.equal(1);
@@ -357,10 +358,10 @@ describe('login controller', () => {
     });
 
     it('should login the user when token is valid', () => {
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
-      sinon.stub(users, 'getUserByToken').resolves('userId');
-      sinon.stub(users, 'resetPassword').resolves({ user: 'user_name', password: 'secret' });
-      sinon.stub(users, 'deactivateTokenLogin').resolves();
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'getUserByToken').resolves('userId');
+      sinon.stub(tokenLogin, 'resetPassword').resolves({ user: 'user_name', password: 'secret' });
+      sinon.stub(tokenLogin, 'deactivateTokenLogin').resolves();
       sinon.stub(request, 'post').resolves({ statusCode: 200, headers: { 'set-cookie': [ 'AuthSession=abc;' ] } });
       sinon.stub(res, 'status').returns(res);
       sinon.stub(res, 'send').returns(res);
@@ -375,11 +376,11 @@ describe('login controller', () => {
         chai.expect(auth.getUserCtx.callCount).to.equal(2);
         chai.expect(auth.getUserCtx.args[0]).to.deep.equal([req]);
         chai.expect(auth.getUserCtx.args[1]).to.deep.equal([{ headers: { 'Cookie': 'AuthSession=abc;' } }]);
-        chai.expect(users.getUserByToken.callCount).to.equal(1);
-        chai.expect(users.resetPassword.callCount).to.equal(1);
-        chai.expect(users.resetPassword.args[0]).to.deep.equal(['userId']);
-        chai.expect(users.deactivateTokenLogin.callCount).to.equal(1);
-        chai.expect(users.deactivateTokenLogin.args[0]).to.deep.equal(['userId']);
+        chai.expect(tokenLogin.getUserByToken.callCount).to.equal(1);
+        chai.expect(tokenLogin.resetPassword.callCount).to.equal(1);
+        chai.expect(tokenLogin.resetPassword.args[0]).to.deep.equal(['userId']);
+        chai.expect(tokenLogin.deactivateTokenLogin.callCount).to.equal(1);
+        chai.expect(tokenLogin.deactivateTokenLogin.args[0]).to.deep.equal(['userId']);
         chai.expect(res.cookie.callCount).to.equal(3);
         chai.expect(res.cookie.args[0].slice(0, 2)).to.deep.equal(['AuthSession', 'abc']);
         chai.expect(res.cookie.args[1].slice(0, 2)).to.deep.equal(['userCtx', JSON.stringify(userCtx) ]);
@@ -392,10 +393,10 @@ describe('login controller', () => {
     });
 
     it('should retry logging in when login fails', () => {
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
-      sinon.stub(users, 'getUserByToken').resolves('userId');
-      sinon.stub(users, 'resetPassword').resolves({ user: 'user_name', password: 'secret' });
-      sinon.stub(users, 'deactivateTokenLogin').resolves();
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'getUserByToken').resolves('userId');
+      sinon.stub(tokenLogin, 'resetPassword').resolves({ user: 'user_name', password: 'secret' });
+      sinon.stub(tokenLogin, 'deactivateTokenLogin').resolves();
       sinon.stub(request, 'post')
         .onCall(0).resolves({ statusCode: 401 })
         .onCall(1).resolves({ statusCode: 401 })
@@ -413,11 +414,11 @@ describe('login controller', () => {
         .onCall(1).resolves(userCtx);
       req.params = { token: 'a', userId: 'b' };
       return controller.tokenPost(req, res).then(() => {
-        chai.expect(users.getUserByToken.callCount).to.equal(1);
-        chai.expect(users.resetPassword.callCount).to.equal(1);
-        chai.expect(users.resetPassword.args[0]).to.deep.equal(['userId']);
-        chai.expect(users.deactivateTokenLogin.callCount).to.equal(1);
-        chai.expect(users.deactivateTokenLogin.args[0]).to.deep.equal(['userId']);
+        chai.expect(tokenLogin.getUserByToken.callCount).to.equal(1);
+        chai.expect(tokenLogin.resetPassword.callCount).to.equal(1);
+        chai.expect(tokenLogin.resetPassword.args[0]).to.deep.equal(['userId']);
+        chai.expect(tokenLogin.deactivateTokenLogin.callCount).to.equal(1);
+        chai.expect(tokenLogin.deactivateTokenLogin.args[0]).to.deep.equal(['userId']);
         chai.expect(request.post.callCount).to.equal(5);
         chai.expect(res.cookie.callCount).to.equal(3);
         chai.expect(res.cookie.args[0].slice(0, 2)).to.deep.equal(['AuthSession', 'cde']);
@@ -432,10 +433,10 @@ describe('login controller', () => {
 
     it('should abandon logging in after retrying 11 times', () => {
       sinon.stub(auth, 'getUserCtx').rejects({ code: 401 });
-      sinon.stub(users, 'validTokenLoginConfig').returns(true);
-      sinon.stub(users, 'getUserByToken').resolves('userId');
-      sinon.stub(users, 'resetPassword').resolves({ user: 'user_name', password: 'secret' });
-      sinon.stub(users, 'deactivateTokenLogin');
+      sinon.stub(tokenLogin, 'validTokenLoginConfig').returns(true);
+      sinon.stub(tokenLogin, 'getUserByToken').resolves('userId');
+      sinon.stub(tokenLogin, 'resetPassword').resolves({ user: 'user_name', password: 'secret' });
+      sinon.stub(tokenLogin, 'deactivateTokenLogin');
       sinon.stub(res, 'status').returns(res);
       sinon.stub(res, 'json');
       sinon.stub(request, 'post').resolves({ statusCode: 401 });
@@ -445,10 +446,10 @@ describe('login controller', () => {
         chai.expect(res.status.args[0]).to.deep.equal([408]);
         chai.expect(res.json.callCount).to.equal(1);
         chai.expect(res.json.args[0]).to.deep.equal([{ error: 'Login failed after 10 retries' }]);
-        chai.expect(users.getUserByToken.callCount).to.equal(1);
-        chai.expect(users.resetPassword.callCount).to.equal(1);
-        chai.expect(users.resetPassword.args[0]).to.deep.equal(['userId']);
-        chai.expect(users.deactivateTokenLogin.callCount).to.equal(0);
+        chai.expect(tokenLogin.getUserByToken.callCount).to.equal(1);
+        chai.expect(tokenLogin.resetPassword.callCount).to.equal(1);
+        chai.expect(tokenLogin.resetPassword.args[0]).to.deep.equal(['userId']);
+        chai.expect(tokenLogin.deactivateTokenLogin.callCount).to.equal(0);
         chai.expect(request.post.callCount).to.equal(11);
       });
     });
