+++ ddocs/medic/_attachments/translations/messages-en.properties
@@ -412,6 +412,7 @@ configuration.enable.token.login.enabled.expired = SMS login is enabled for this
 configuration.enable.token.login.enabled.inactive = SMS login is enabled for this user. The link was accessed on {{date}}.
 configuration.enable.token.login.disable = Disable login via SMS.  
 configuration.enable.token.login.refresh = Regenerate and re-send login SMS.
+configuration.enable.token.login.no.modify = Make no changes to SMS login.
 configuration.enable.token.login.refresh.help = Disabling or regenerating the login SMS will result in changing the user's password, which will result in the user being logged out. 
 configuration.messagetest = Message test
 configuration.permission = Permission
